#84: Feature Styler2 (http://github.com/Polymap4/polymap4-p4/issues/issue/84)
  - more type information about the actual target type of a style property: StylePropertyValue<T>
  - FeatureStyle: one UnitOfWork per instance
  - StyleRepository: central cache of serialized versions
  - TypeValueInitialize is deprecated

+++ plugins/org.polymap.core.style/src/org/polymap/core/style/serialize/sld/PointStyleSerializer.java
@@ -40,7 +40,7 @@ public class PointStyleSerializer
 
     @Override
     public void doSerialize( PointStyle style ) {
-        setValue( style.strokeWidth.get(), (PointSymbolizerDescriptor sd, Integer value) -> sd.strokeWidth.set( value ) );
+        setValue( style.strokeWidth.get(), (PointSymbolizerDescriptor sd, Double value) -> sd.strokeWidth.set( value ) );
         setValue( style.strokeOpacity.get(), (PointSymbolizerDescriptor sd, Double value) -> sd.strokeOpacity.set( value ) );
         setValue( style.strokeColor.get(), (PointSymbolizerDescriptor sd, Color value) -> sd.strokeColor.set( value ) );
         setValue( style.fillColor.get(), (PointSymbolizerDescriptor sd, Color value) -> sd.fillColor.set( value ) );
